Started implementing fish
Fish now swim around and can detect the bobber. The fish will swim to the bobber and stop there.

Next to do is to make the state change start and collect the fish when the player reels it in.

I'm also thinking about making an interface for fish, some kind of scriptable object that will be extendable and that I can make different types of fish for.