Have a good loop for fishing working.
Finally got the animation event to call correctly. Now I just need to start working on a basic inventory to add fish to. And I'm thinking I may define areas for fish to start inside of.

Could be a good idea to make different fish type SO's that can populate info on the Fish prefab.